unit test for chat template
